backend/requirement.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
multidict==6.0.2
psycopg2-binary==2.9.5
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
